// src/types/cooperative.type.ts
export type TCooperativeInfo = {
  id?: number;
  cooperativeCode?: string;
  coopsFullNameNep?: string;
  coopsFullNameEng?: string;
  registerDate?: string;
  registerYear?: string;
  fiscalYearId?: number;
  registeredFiscalYear?: string;
  contactMobilePhone?: string;
  contactOfficePhone?: string;
  webUrl?: string;
  contactPerson?: string;
  contactPhone?: string;
  contactEmail?: string;
  panNo?: string;
  registrationNo?: string;
  classificationOfCooperative?: string;
  workingArea?: string;
  anyBranchOffice?: boolean;
  numberOfBranch?: number;
  folderName?: string;
  registeredFederation?: boolean;
  status?: string;
  fiscalYear?: TFiscalYear;
  createdBy?: string;
  createdDate?: string;
  updatedBy?: string;
  updatedDate?: string;
};

export type TDocument = {
  id: number | string;
  name: string;
  type: "image" | "pdf";
  url: string;
};

export type TCoopAddressInfo = {
  id?: number;
  cooperativeInfoId?: number;
  province?: string;
  district?: string;
  localLevel?: string;
  wardNo?: number;
  houseNo?: string;
  tole?: string;
};

export type TCoopStrenght = {
  createdBy?: string;
  createdDate?: string;
  updatedBy?: string;
  updatedDate?: string;
  id?: number;
  coopAddressInfoId?: number;
  coopAddressInfo?: {
    createdBy?: string;
    createdDate?: string;
    updatedBy?: string;
    updatedDate?: string;
    id?: number;
    cooperativeInfoId?: number;
    cooperativeInfo: {
      createdBy?: string;
      createdDate?: string;
      updatedBy?: string;
      updatedDate?: string;
      id?: number;
      cooperativeCode?: string;
      coopsFullNameNep?: string;
      coopsFullNameEng?: string;
      registerDate?: string;
      registerYear?: string;
      fiscalYearId?: number;
      fiscalYear?: {
        createdBy?: string;
        createdDate?: string;
        updatedBy?: string;
        updatedDate?: string;
        id?: number;
        yearNepali?: string;
        yearEnglish?: string;
        activeFiscalYear?: boolean;
        index?: number;
      };
      registeredFiscalYear?: string;
      contactMobilePhone?: string;
      contactOfficePhone?: string;
      webUrl?: string;
      contactPerson?: string;
      contactPhone?: string;
      contactEmail?: string;
      panNo?: string;
      registrationNo?: string;
      classificationOfCooperative?: string;
      workingArea?: string;
      anyBranchOffice?: boolean;
      numberOfBranch?: number;
      status?: string;
      states?: number[];
      folderName?: string;
      registeredFederation?: boolean;
    };
    province?: string;
    district?: string;
    localLevel?: string;
    wardNo?: number;
    houseNo?: string;
    tole?: string;
  };
  cooperativeInfoId?: number;
  cooperativeInfo?: {
    createdBy?: string;
    createdDate?: string;
    updatedBy?: string;
    updatedDate?: string;
    id?: number;
    cooperativeCode?: string;
    coopsFullNameNep?: string;
    coopsFullNameEng?: string;
    registerDate?: string;
    registerYear?: string;
    fiscalYearId?: number;
    fiscalYear: {
      createdBy?: string;
      createdDate?: string;
      updatedBy?: string;
      updatedDate?: string;
      id?: number;
      yearNepali?: string;
      yearEnglish?: string;
      activeFiscalYear: boolean;
      index?: number;
    };
    registeredFiscalYear?: string;
    contactMobilePhone?: string;
    contactOfficePhone?: string;
    webUrl?: string;
    contactPerson?: string;
    contactPhone?: string;
    contactEmail?: string;
    panNo?: string;
    registrationNo?: string;
    classificationOfCooperative?: string;
    workingArea?: string;
    anyBranchOffice?: boolean;
    numberOfBranch?: number;
    status?: string;
    states?: number[];
    folderName?: string;
    registeredFederation: boolean;
  };
  totalShareAmount?: number;
  numberOfTotalShare?: number;
  noOfTotalMaleMember?: number;
  noOfTotalFemaleMember?: number;
  totalSavingAmount?: number;
  totalLiabilities?: number;
  loanInOperation?: number;
  totalAmountOfInstitudeAssets?: number;
  verificationStatus?: boolean;
};

export type TCommitteeMember = {
  id?: number;
  cooperativeInfoId?: number;
  committeeType?: string;
  commiteeFormulationAssembly?: string;
  position?: string;
  firstName?: string;
  middleName?: string;
  lastName?: string;
  gender?: string;
  ethnicity?: string;
  email?: string;
  phoneNumber?: string;
  panNo?: string;
  nid?: string;
  province?: string;
  district?: string;
  localLevel?: string;
  wardNo?: number;
  houseNumber?: string;
  tole?: string;
  panCard?: string;
  nidCard?: string;
  highestEducation?: string;
};
// ...
export type TExecutiveMember = {
  id?: number;
  cooperativeInfoId?: number;
  executiveMemberType?: string;
  committeeFormulationAssembly?: string;
  position?: string;
  firstName?: string;
  middleName?: string;
  lastName?: string;
  gender?: string;
  ethnicity?: string;
  email?: string;
  phoneNumber?: string;
  panNo?: string;
  nid?: string;
  province?: string;
  district?: string;
  localLevel?: string;
  wardNo?: number;
  houseNumber?: string;
  tole?: string;
  panCard?: string;
  nidCard?: string;
  highestLevelEducation?: string;
};

export type TDocumentsInfo = {
  id?: number;
  cooperativeInfoId?: number;
  businessRegisterCertificate?: string;
  panRegisterCertificate?: string;
  otherRegisterCertificate?: string;
  approvaLbylaw?: string;
  feasibilityReport?: string;
  remark?: string;
  createdBy?: string;
  createdDate?: string;
  updatedBy?: string;
  updatedDate?: string;
};
// ...
export type TUploadAODoc = {
  id?: number;
  cooperativeInfoId?: number;
  cooperativeInfo?: TCooperativeInfo;
  fiscalYearId?: number;
  fiscalYear?: TFiscalYear;
  documentType?: string;
  documentUpload?: string;
  auditReport?: string;
  taxClearance?: string;
  boardMeetingDecision?: string;
  annualMeetingAttendence?: string;
  electionRelatedDoc?: string;
  otherDoc?: string;
  remarks?: string;
  createdBy?: string;
  createdDate?: string;
  updatedBy?: string;
  updatedDate?: string;
};

export type TCooperativeDetailInfo = {
  committeeDetails: TCommitteeMember[];
  coopAddressInfo: TCoopAddressInfo;
  coopBranchRegisters: unknown;
  coopStrength: TCoopStrenght;
  cooperativeInfo: TCooperativeInfo;
  executiveDetails: TExecutiveMember[];
  uploadAODoc: TUploadAODoc;
  uploadRegistrationDoc: TDocumentsInfo;
};

export type TGetCooperativeDetailInfoByCoopIdResponse = {
  data: TCooperativeDetailInfo;
  page: number;
  pageSize: number;
  totalPages: number;
  totalRecords: number;
};